feat: Allow to set slack channel per job dynamically
meta set build.meta.notification.slack.{jobname}.channels can be used to overwrite the yaml defined slack channel.
Overwrites are specific to the current job.
It can be a comma separated string for multiple channels. This will only work to overwrite a yaml set slack channel. It is not a replacement.

This is useful when code inside a SD step decides who should be notified.